entailment/requirements.txt
scikit-learn
seqeval
psutil
sacrebleu
rouge-score
pytorch-lightning==0.8.5
elasticsearch
pandas
nlp